api/requirements.txt
Flask
flask-cors
SpeechRecognition
python-dotenv
google-generativeai
